backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
websockets==15.0.1
paramiko==4.0.0
pandas==2.3.3
openpyxl==3.1.5
pyxlsb==1.0.10
xlrd==2.0.2
xlsxwriter==3.2.9
scikit-learn==1.8.0
chardet==5.2.0
tqdm==4.67.1
python-dotenv==1.2.1
aiofiles==25.1.0